requirements.txt
Django==4.2.3
pytrends==4.8.0
pandas==2.0.3
statsmodels==0.14.0
plotly==5.14.1
numpy==1.25.0
google-generativeai==0.1.0
markdown==3.4.4
